Use a single thread for now
I am running into some issues, and am not sure if it the application or a gem used during testing. When running the tests, the process sometimes hangs when selenium webdriver tries to close Firefox. This is the only obvious issue I've noticed.

I did notice that my test suite ran much faster once I set Puma to run with a single thread, which I think suggests there were race conditions (I'm not sure if that is what you call it, but I'm thinking of multiple threads fighting for the same memory resources)?