add ExtractValue concern
Provides #value method to models. This method is intended to be used to
compare records by value, rather than by identity. It returns a hash of
the attribute names and values (for the whitelisted attributes).

In the a couple commits ahead of this, you’ll see that I use this to
determine if records
should be considered “blank”.