try having selenium use Chrome instead of Firefox
Firefox is still hanging when it comes time for it to close at the end of the spec suite => making Puma run with a single thread didn't help.

I've googled a little bit, and it sounds like maybe this is a general issue with Firefox  and Selenium. For now, I'll try switching to Chrome. If that doesn't work, I'll try Poltergeist or the thing that Thoughtbot made. If it does work consistently, I'll try having Puma use multiple threads again.