Fixed issue with RSpec hanging
I think... it seems that the issue was that I was trying to override the base strategy of transaction cleanup in a separate file, and that doesn't work. I moved it all to the same file (now all in the `database_cleaner_config.rb` file). I didn't do this at first, because I didn't want to have to load Capybara for non-feature specs. But that line of code that references Capybara doesn't execute until after Capybara is loaded (via `feature_helper.rb`), so it all works out.