Implement functionality to add climbs via the gym show page
Also…
- JS flash handler needs to be set on every page update, not just when
the page changes.
- Climb type uses a radio button group - if we use this elsewhere, we
should make a mixing for radio button groups.
- Added a show partial for gym sections. This displays all of the
climbs for that section, and it is where the button to create a new
climb is (so that the climb can automatically be associated with that
section). Later, we’ll want to only show “active” climbs in this
partial.
- A climb must belong to a section, and this is enforced via an
ActiveRecord validation and a DB constraint.
- A climb must have a type (Boulder or Route), and this is enforced via
an ActiveRecord validation and a DB constraint.
- Fixed the Guardfile to use more patterns rather than listing out a
bunch of strings.
- Make sure we check for use of strong parameters in
GymsController#update
- The views that I created still need to be styled.