Create and enforce `GymPolicy`
The `:new`, `:create`, `:edit`, and `:update` actions should check for authorization.  Links to these actions are shown only if the user is authorized to perform them.

The policy specs will verify the policy is behaving as expected, the controller specs will verify the policies are used where expected. The controller specs stub out the policies to return false because all we want to check is that the policy is *used*, the policy specs handle the unit testing. At some point it would be nice to have a more concise way to check that controllers apply the policies.