Modify policies to look at current role only
It isn't really a problem for an admin to create a gym or climb when their current role is set to a non-admin value, however we are currently using the policies to determine whether or not to show links, and I want the admin to see the site as a regular users would see it when their current role is switched to something else (e.g. athlete).

This might be overloading the purpose of the policy class because it is now responsible for (a) whether an action should be allowed and (b) whether the option to take an action should be presented to the user. But I think that in general, these two should be the same thing. We can change it later if we need to.