Create seed data for grade systems
I am planning on re-doing this in the next commit. I'd like to treat the seed data like the database schema, using migrations to define how the seed data should change, and keeping a file (or rather files) that define the current state of the seed data (equivalent to the schema.rb file for database schema).