Start connecting the new Grade and GradeSystem models to the climbs / sections / gyms
This is a broken commit and will be fixed in the next commit. We need to rework the way that the list of grades is populated on the form for a new climb. Previously, we could just get the choices from the Route / Boulder enum, but now we need to know what gym the climb is being created for because each gym can have their own grade system. Currently, the new action for climbs doesn't know what the gym is. I think I'm going to nest the section resource under gyms and climbs under sections to solve this.