Nest Gym :new and :create under Section
Still don't have all tests passing. This commit changes the routes for creating a new climb to be nested under the Section resource because we need to know wha the Section (or rather, the Gym) is so that we can show the correct list of grades for the climb.

There were problems with Bullet giving false positives for N+1 queries. When you call those methods in the console, you see that these aren't actually N+1 issues... so I whitelisted them.

Deleted the specs that tested for the proper error handling when a section wasn't specified for the climb because that isn't possible anymore (the section is needed for the route).