Refill seed_migrations table if it isn't up to date with current version
Do this so that if we run something like `rake db:reset` and the seed_migratios table gets dropped, we don't re-run the old seed migrations. We could have just adopted the practice that all migrations need to be able to be run multiple times (and that is probably still a good idea), but this approach is more robust.

This code will need to be refactored...