Create Employment model
This is a join model for the many to many relationship between SetterStory (and eventually ManagerStory, etc.) and Gym.

We should change EmployeesController over to EmploymentsController#index, and just modify the route name.

Still need to do validation and db constraints on employments, and look at how the response goes when the employment is invalid.