Change the Form objects design to accomodate ActiveRecord::Validations
In EmploymentForm, there are validations that are specific to the form,
not to any particular model. For this reason, I wanted to mixin
ActiveRecord::Validations. I had also seen that there were some other
modules I could mixin instead of delegating specific methods to object
in EmploymentForm. The changes I made caused fields to be named
differently, so I had to change quite a bit of selectors in the
stylesheets and javascript, the params filtering methods, and some
stuff in the specs and page objects.

Also in this commit…
- fix border on employee list table
- make EmploymentForm handle invalid submissions
- increased max wait time yet again in the reset feature spec…