Fix case where current employee is given a new role
Needed to add the role to the user record / create the role story
record. Also needed to rework the Javascript so that the existing row
in the table is modified, rather than adding a new row. To accommodate
this, the usage of the Employee model was changed - now it must be
initialized with an email and a gym id because the roles are always
calculated based on this information. This ensures that roles always
returns *all* roles that the user has, scoped to the given gym.
Previously Employee was just a Struct like class, and was only used in
EmployeeList. Now EmployeeList just uses OpenStruct so that each item
in the list can implement the Employee interface.

Changes were made to BaseForm#save! so that the error that is raised
will include details about why the record was invalid.