Try making the selectors more specific
For some reason, on Semaphore (only, no on my local machine), it is
showing that there are multiple matches for the selector I had before.
I don’t see any reason why this would be different on Semaphore,
especially since this shouldn’t be a time dependent issue. Maybe this
will help?