Only check id attribute
For some reason, on Semaphore only, the created_at and updated_at
attributes are different in memory than they are in the db, but only by
an insignificant amount. I don’t understand how this happens, but it
also isn’t important that they are exactly the same, so we’ll just
check the id’s.