Enforce that `Role` records can only be created via seed data
This prevents roles from being created by end users (e.g. if someone were to craft their own form submission). Right now the consequences of this record being created are pretty low, but it could be more of a problem later. Plus, it was interesting to implement.

In order to support this, we needed to be able to handle polymorphic associations in `TableDependencyGraph`, because the roles table has this type of association (`resource`). `TableDependencyGraph` should be refactored because it is hard to follow right now.

A few minor changes were made to `Seedster` so that we can ask it if it is currently migrating (that is the only time we'll allow records to be added to a seeds only table).

On an unrelated note, I started scoping the test classes with anonymous modules so that if I use `Post` as a test class in two separate, unrelated specs, they don't interfere with each other (e.g. see `active_record_helpers_spec` and `extract_value_spec`).