Use custom validation for role whitelist
This means that the `Employment.roles` list can change while the application is running (e.g. if we eventually add functionality for end users to define roles). More important for right now is that it means it doesn't matter that `app/forms/employment_form.rb` is loaded before the role story models.