Guard the `ClimbLogsController`
The actions should only be performed for athletes (there are application errors if `current_user` doesn't have an `athlete_story`).

Once we have more athlete specific controllers, it would be nice to create a parent class that defines this before action, and then have all the controllers in the `athletes` folder inherit from this.