Implement `RoleStories.for_user`
This method allows us to get all role story records for a particular user - if a user is a manager and a setter, it will return both the `ManagerStory` and `SetterStory` records.

This allows us to clean up `Employment#for_user` because we can move the logic for finding the role stories to the `RoleStories` module, where it belongs. Now, all we have to do is join on all role story tables, and then merge on the `RoleStories.for_user` query.